+++ lib/src/controls/inputs/buttons/base.dart
@@ -161,10 +161,8 @@ class _BaseButtonState extends State<BaseButton> {
         final resolvedPadding = resolve<EdgeInsetsGeometry?>(
                 (ButtonStyle? style) => style?.padding) ??
             EdgeInsets.zero;
-        final resolvedBorder =
-            resolve<BorderSide?>((ButtonStyle? style) => style?.border);
         final resolvedShape =
-            resolve<OutlinedBorder?>((ButtonStyle? style) => style?.shape) ??
+            resolve<ShapeBorder?>((ButtonStyle? style) => style?.shape) ??
                 const RoundedRectangleBorder();
 
         final padding = resolvedPadding
@@ -187,7 +185,7 @@ class _BaseButtonState extends State<BaseButton> {
             duration: FluentTheme.of(context).fasterAnimationDuration,
             curve: FluentTheme.of(context).animationCurve,
             decoration: ShapeDecoration(
-              shape: resolvedShape.copyWith(side: resolvedBorder),
+              shape: resolvedShape,
               color: resolvedBackgroundColor,
             ),
             padding: padding,