+++ lib/src/controls/inputs/buttons/filled_button.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:fluent_ui/fluent_ui.dart';
 
 /// A colored button.
@@ -43,6 +45,9 @@ class FilledButton extends Button {
       foregroundColor: ButtonState.resolveWith(
         (states) => foregroundColor(theme, states),
       ),
+      shape: ButtonState.resolveWith((states) {
+        return shapeBorder(theme, states);
+      }),
     );
 
     return super.defaultStyleOf(context).merge(def) ?? def;
@@ -74,5 +79,29 @@ class FilledButton extends Button {
       return res.textOnAccentFillColorDisabled;
     }
     return res.textOnAccentFillColorPrimary;
+  }
+
+  static ShapeBorder shapeBorder(
+      FluentThemeData theme, Set<ButtonStates> states) {
+    return states.isPressing || states.isDisabled
+        ? RoundedRectangleBorder(
+            side: BorderSide(
+              color: theme.resources.controlFillColorTransparent,
+            ),
+            borderRadius: BorderRadius.circular(4.0),
+          )
+        : RoundedRectangleGradientBorder(
+            gradient: LinearGradient(
+              begin: const Alignment(0.0, -2),
+              end: Alignment.bottomCenter,
+              colors: [
+                theme.resources.controlStrokeColorOnAccentSecondary,
+                theme.resources.controlStrokeColorOnAccentDefault,
+              ],
+              stops: const [0.33, 1.0],
+              transform: const GradientRotation(pi),
+            ),
+            borderRadius: BorderRadius.circular(4.0),
+          );
   }
 }